pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-agents"
version = "0.1.0"
description = "Building and testing different AI agents and agentic system architectures"
authors = [
    {name = "Cheyanne Allred-Lopez", email = "cea33@students.uwf.edu"}
]
requires-python = ">=3.11, <3.13"
dependencies = [
    "langchain~=1.2.0",
    "langgraph~=1.0.5",
    "langchain-google-genai~=4.0",
    "python-dotenv>=1.2.1", 
    "pydantic>=2.0.0",
    "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "pylint>=2.12.0",
    "mypy>=0.950",
    "types-PyYAML>=6.0.0",
    "pre-commit>=2.20.0",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.pylint]
max-line-length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
